requirements.txt
requests~=2.26.0
tweepy~=3.10.0
SQLAlchemy~=1.4.23
MetaTrader5~=5.0.35
matplotlib~=3.4.3
pandas~=1.3.2
aiohttp~=3.7.4.post0
beautifulsoup4~=4.9.3
elasticsearch~=7.14.1
geopy~=2.2.0
googletransx~=2.4.2
dataclasses~=0.6
pymongo~=3.12.0
pyquery~=1.4.3
python-dotenv~=0.19.0
dnspython
sql_formatter
pymsql
Unidecode~=1.3.2
selenium~=3.141.0